requirements.txt
fastapi
uvicorn
sqlmodel
sqladmin
pyyaml
python-multipart
